XD-3385 Added CloudConfiguration via @Import for "cloud" profiles
- this allows us to handle "lattice" and "!lattice" profiles only for "cloud"
